+++ web/src/Components/HistoryTasks.js
@@ -1,0 +1,7 @@
+import React, {useEffect, useRef, useState} from "react";
+
+function HistoryTasks() {
+  return <div>History Tasks page - to be implemented</div>;
+}
+
+export default HistoryTasks;
